refactor(mcp): use dynamic import for subcommand
