+++ packages/opencode/src/index.ts
@@ -18,7 +18,7 @@ import { ServeCommand } from "./cli/cmd/serve"
 import { Filesystem } from "@/util/filesystem"
 import { DebugCommand } from "./cli/cmd/debug/command"
 import { StatsCommand } from "./cli/cmd/stats"
-import { McpCommand } from "./cli/cmd/mcp"
+import { McpCommand } from "./cli/cmd/mcp/command"
 import { GithubCommand } from "./cli/cmd/github/command"
 import { ExportCommand } from "./cli/cmd/export/command"
 import { ImportCommand } from "./cli/cmd/import/command"
